# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  code: string;
}

export interface UserRegistrationInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}

export interface UserRegistrationResult {
  userId: string;
  name: string;
  hasSelection: boolean;
  /** @nullable */
  selectedNumber: number | null;
  availableCount: number;
}

export interface NumberItem {
  /**
     * @minimum 1
     * @maximum 14
     */
  number: number;
  available: boolean;
}

export interface NumbersResponse {
  numbers: NumberItem[];
  /** @nullable */
  selectedNumber: number | null;
  availableCount: number;
}

export interface NumberSelectionInput {
  /**
     * @minimum 1
     * @maximum 14
     */
  number: number;
}

export interface SelectionResult {
  selectedNumber: number;
  selectedAt: string;
}

export interface MeResponse {
  authenticated: boolean;
  /** @nullable */
  name: string | null;
  /** @nullable */
  selectedNumber: number | null;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminLoginResult {
  authenticated: boolean;
}

export type AdminAssignmentStatus = typeof AdminAssignmentStatus[keyof typeof AdminAssignmentStatus];


export const AdminAssignmentStatus = {
  selected: 'selected',
  available: 'available',
} as const;

export interface AdminAssignment {
  number: number;
  /** @nullable */
  name: string | null;
  status: AdminAssignmentStatus;
  /** @nullable */
  selectedAt: string | null;
}

export interface AdminDashboard {
  totalNumbers: number;
  selected: number;
  remaining: number;
  assignments: AdminAssignment[];
}