# lib/api-zod/src/generated/types/adminAssignment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAssignmentStatus } from './adminAssignmentStatus';

export interface AdminAssignment {
  number: number;
  /** @nullable */
  name: string | null;
  status: AdminAssignmentStatus;
  /** @nullable */
  selectedAt: Date | null;
}