# lib/api-zod/src/generated/types/adminDashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAssignment } from './adminAssignment';

export interface AdminDashboard {
  totalNumbers: number;
  selected: number;
  remaining: number;
  assignments: AdminAssignment[];
}