# lib/api-zod/src/generated/types/meResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MeResponse {
  authenticated: boolean;
  /** @nullable */
  name: string | null;
  /** @nullable */
  selectedNumber: number | null;
}