# lib/api-zod/src/generated/types/numberItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NumberItem {
  /**
     * @minimum 1
     * @maximum 14
     */
  number: number;
  available: boolean;
}