# lib/api-zod/src/generated/types/numberSelectionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NumberSelectionInput {
  /**
     * @minimum 1
     * @maximum 14
     */
  number: number;
}