# lib/api-zod/src/generated/types/numbersResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NumberItem } from './numberItem';

export interface NumbersResponse {
  numbers: NumberItem[];
  /** @nullable */
  selectedNumber: number | null;
  availableCount: number;
}