# lib/api-zod/src/generated/types/selectionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SelectionResult {
  selectedNumber: number;
  selectedAt: Date;
}