# lib/api-zod/src/generated/types/userRegistrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserRegistrationInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}