# lib/api-zod/src/generated/types/userRegistrationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserRegistrationResult {
  userId: string;
  name: string;
  hasSelection: boolean;
  /** @nullable */
  selectedNumber: number | null;
  availableCount: number;
}